+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - N/A
 
 ### Added
-- 
+- module find_sap_apps for linux.
 
 ### Security
 - N/A